chore: move sprites config types to utils-cli chore: add defineConfig chore: add package exports + dts
