+++ packages/utils-cli/build.config.ts
@@ -2,11 +2,19 @@ import { defineBuildConfig } from 'unbuild';
 
 export default defineBuildConfig({
 	outDir: './lib',
-	entries: [{ input: 'src/svg-css-sprite/cli', name: 'svg-css-sprite/cli' }],
+	entries: [
+		{
+			input: 'src/svg-css-sprite/index',
+			name: 'svg-css-sprite/index',
+		},
+		{
+			input: 'src/svg-css-sprite/cli',
+			name: 'svg-css-sprite/cli',
+		},
+	],
 	clean: true,
-	declaration: false,
-	failOnWarn: false,
+	declaration: true,
 	rollup: {
-		emitCJS: false,
+		emitCJS: true,
 	},
 });
