+++ packages/utils-cli/src/svg-css-sprite/index.ts
@@ -1,0 +1,7 @@
+import type { SpritesConfiguration } from './types';
+
+export * from './types';
+
+export default function defineConfig(config: SpritesConfiguration) {
+	return config;
+}
