+++ packages/utils-cli/test-cli/custom.ts
@@ -1,7 +1,7 @@
-import { SpritesConfiguration } from '@iconify/utils/lib/svg-css-sprite/types';
 import { createLoadCollectionFromFSAsyncIterator } from '@iconify/utils/lib/svg-css-sprite/create-node-sprite';
+import defineConfig from '../lib/svg-css-sprite';
 
-export default <SpritesConfiguration>{
+export default defineConfig({
 	sprites: [
 		{
 			name: 'flat-color-icons-custom',
@@ -14,4 +14,4 @@ export default <SpritesConfiguration>{
 			outdir: './',
 		},
 	],
-};
+});
