+++ packages/utils-cli/test-cli/svg-css-sprite.config.ts
@@ -1,17 +1,15 @@
-import { SpritesConfiguration } from '@iconify/utils/lib/svg-css-sprite/types';
+import type { SpritesConfiguration } from '../lib/svg-css-sprite';
 import { createLoadCollectionFromFSAsyncIterator } from '@iconify/utils/lib/svg-css-sprite/create-node-sprite';
 
-export default <SpritesConfiguration>{
-	sprites: [
-		{
-			name: 'flat-color-icons',
-			collection: createLoadCollectionFromFSAsyncIterator(
-				'flat-color-icons',
-				{
-					include: ['about', 'accept-database'],
-				}
-			),
-			outdir: './',
-		},
-	],
-};
+export default {
+	sprites: {
+		name: 'flat-color-icons',
+		collection: createLoadCollectionFromFSAsyncIterator(
+			'flat-color-icons',
+			{
+				include: ['about', 'accept-database'],
+			}
+		),
+		outdir: './',
+	},
+} satisfies SpritesConfiguration;
