+++ packages/utils/src/index.ts
@@ -128,8 +128,6 @@ export {
 export type {
 	AsyncSpriteIcons,
 	AsyncSpriteIconsFactory,
-	SpriteConfiguration,
-	SpritesConfiguration,
 	SpriteIcon,
 	SpriteIcons,
 	SpriteCollection,
