+++ packages/utils/src/svg-css-sprite/types.ts
@@ -31,14 +31,3 @@ export type SpriteCollection =
 	| SpriteIcon[]
 	| AsyncSpriteIcons
 	| AsyncSpriteIconsFactory;
-
-export interface SpriteConfiguration {
-	name: string;
-	collection: SpriteCollection | SpriteCollection[];
-	outdir?: string;
-	mapIconName?: (icon: string, collection?: string) => string;
-}
-
-export interface SpritesConfiguration {
-	sprites: SpriteConfiguration | SpriteConfiguration[];
-}
